+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="com.delaroystudios.CardView_GoogleApp.CardDemoActivity"
+            android:name="com.delaroystudios.CardView_GoogleApp.GoogleNowActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
