--- app/src/main/java/com/delaroystudios/CardView_GoogleApp/CardDemoActivity.java
@@ -1,27 +1,0 @@
-package com.delaroystudios.CardView_GoogleApp;
-
-import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import com.google.android.youtube.player.YouTubeBaseActivity;
-
-public class CardDemoActivity extends YouTubeBaseActivity {
-
-    private RecyclerView recyclerView;
-    private RecyclerView.LayoutManager layoutManager;
-    private RecyclerView.Adapter adapter;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_card_demo);
-
-        recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
-        layoutManager = new LinearLayoutManager(this);
-        adapter = new RecyclerAdapter();
-
-        recyclerView.setLayoutManager(layoutManager);
-        recyclerView.setAdapter(adapter);
-
-    }
-}
