--- app/src/main/res/layout/activity_card_demo.xml
@@ -1,27 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#AA0000"
-    android:orientation="vertical"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    tools:context=".CardDemoActivity">
-
-    <TextClock
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:gravity="end"
-        android:textSize="45dp" />
-
-    <CalendarView
-        android:layout_width="wrap_content"
-        android:layout_height="150dp"/>
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-</LinearLayout>
