+++ app/src/main/res/layout/youtube_activity.xml
@@ -4,15 +4,14 @@
     android:id="@+id/card_view"
     android:layout_width="match_parent"
     android:layout_height="230dp"
-    android:gravity="center"
     android:layout_margin="5dp"
     android:clickable="true"
     android:foreground="?selectableItemBackground"
+    android:gravity="center"
     card_view:cardBackgroundColor="#AA000000"
     card_view:cardCornerRadius="12dp"
     card_view:cardElevation="5dp"
     card_view:contentPadding="4dp">
-
 
     <RelativeLayout
         android:layout_width="match_parent"
@@ -42,5 +41,6 @@
             android:layout_toEndOf="@+id/play_kanye"
             android:layout_toRightOf="@+id/play_kanye"
             android:gravity="center" />
+
     </RelativeLayout>
 </android.support.v7.widget.CardView>
